Modified the helper's API
Implemented lwc_copy_pgroot and lwc_cow_pgroot, both relying on a newly implemented page_walker.
For the moment, since we do not have kernel and user space, we cannot make the current memory mapping COW.
I therefore rely on a hack that generates three copies and is inefficient.
Next:
Update the page count for shared pages.
Correct the hack above with a better solution.