Saving progress.
Readonly does not reflect changes in the parent for the moment.
This is due to the use of mmap, that does not take memory from the page pool.
As a result, we cannot count the number of references for this page and make sure the child shares it with the parent.
I'll try to uncow the parent too and make both pages point to the same address instead.